+++ resources/views/layouts/app.blade.php
@@ -1,0 +1,17 @@
+<!DOCTYPE html>
+<html lang="{{ str_replace('_', '-', app()->getLocale()) }}">
+
+<head>
+    <meta charset="utf-8">
+    <title>{{ config('app.name', 'Laravel') }}</title>
+</head>
+
+<body>
+    <div id="app">
+        <main class="py-4">
+            @yield('content')
+        </main>
+    </div>
+</body>
+
+</html>